removal of bigwriteaof, info, lru commands and migration of bitop
