+++ internal/eval/store_eval.go
@@ -6505,3 +6505,159 @@ func evalTYPE(args []string, store *dstore.Store) *EvalResponse {
 		Error:  nil,
 	}
 }
+
+// BITOP <AND | OR | XOR | NOT> destkey key [key ...]
+func evalBITOP(args []string, store *dstore.Store) *EvalResponse {
+	operation, destKey := args[0], args[1]
+	operation = strings.ToUpper(operation)
+
+	// get all the keys
+	keys := args[2:]
+
+	// validation of commands
+	// if operation is not from enums, then error out
+	if !(operation == AND || operation == OR || operation == XOR || operation == NOT) {
+		return makeEvalError(diceerrors.ErrSyntax)
+	}
+
+	if operation == NOT {
+		if len(keys) != 1 {
+			return makeEvalError(diceerrors.ErrGeneral("BITOP NOT must be called with a single source key"))
+		}
+		key := keys[0]
+		obj := store.Get(key)
+		if obj == nil {
+			return makeEvalResult(clientio.IntegerZero)
+		}
+
+		var value []byte
+
+		switch oType, _ := object.ExtractTypeEncoding(obj); oType {
+		case object.ObjTypeByteArray:
+			byteArray := obj.Value.(*ByteArray)
+			byteArrayObject := *byteArray
+			value = byteArrayObject.data
+			// perform the operation
+			result := make([]byte, len(value))
+			for i := 0; i < len(value); i++ {
+				result[i] = ^value[i]
+			}
+
+			// initialize result with byteArray
+			operationResult := NewByteArray(len(result))
+			operationResult.data = result
+			operationResult.Length = int64(len(result))
+
+			// resize the byte array if necessary
+			operationResult.ResizeIfNecessary()
+
+			// create object related to result
+			obj = store.NewObj(operationResult, -1, object.ObjTypeByteArray, object.ObjEncodingByteArray)
+
+			// store the result in destKey
+			store.Put(destKey, obj)
+			return makeEvalResult(len(value))
+
+		case object.ObjTypeString, object.ObjTypeInt:
+			if oType == object.ObjTypeString {
+				value = []byte(obj.Value.(string))
+			} else {
+				value = []byte(strconv.FormatInt(obj.Value.(int64), 10))
+			}
+			// perform the operation
+			result := make([]byte, len(value))
+			for i := 0; i < len(value); i++ {
+				result[i] = ^value[i]
+			}
+			resOType, resOEnc := deduceTypeEncoding(string(result))
+			var storedValue interface{}
+			if resOType == object.ObjTypeInt {
+				storedValue, _ = strconv.ParseInt(string(result), 10, 64)
+			} else {
+				storedValue = string(result)
+			}
+			store.Put(destKey, store.NewObj(storedValue, -1, resOType, resOEnc))
+			return makeEvalResult(len(value))
+
+		default:
+			return makeEvalError(diceerrors.ErrWrongTypeOperation)
+		}
+	}
+	// if operation is AND, OR, XOR
+	values := make([][]byte, len(keys))
+
+	// get the values of all keys
+	for i, key := range keys {
+		obj := store.Get(key)
+		if obj == nil {
+			values[i] = make([]byte, 0)
+		} else {
+			// handle the case when it is byte array
+			switch oType, _ := object.ExtractTypeEncoding(obj); oType {
+			case object.ObjTypeByteArray:
+				byteArray := obj.Value.(*ByteArray)
+				byteArrayObject := *byteArray
+				values[i] = byteArrayObject.data
+			case object.ObjTypeString:
+				value := obj.Value.(string)
+				values[i] = []byte(value)
+			case object.ObjTypeInt:
+				value := strconv.FormatInt(obj.Value.(int64), 10)
+				values[i] = []byte(value)
+			default:
+				return makeEvalError(diceerrors.ErrWrongTypeOperation)
+			}
+		}
+	}
+	// get the length of the largest value
+	maxLength := 0
+	minLength := len(values[0])
+	maxKeyIterator := 0
+	for keyIterator, value := range values {
+		if len(value) > maxLength {
+			maxLength = len(value)
+			maxKeyIterator = keyIterator
+		}
+		minLength = min(minLength, len(value))
+	}
+
+	result := make([]byte, maxLength)
+	if operation == AND {
+		for i := 0; i < maxLength; i++ {
+			result[i] = 0
+			if i < minLength {
+				result[i] = values[maxKeyIterator][i]
+			}
+		}
+	} else {
+		for i := 0; i < maxLength; i++ {
+			result[i] = 0x00
+		}
+	}
+
+	// perform the operation
+	for _, value := range values {
+		for i := 0; i < len(value); i++ {
+			switch operation {
+			case AND:
+				result[i] &= value[i]
+			case OR:
+				result[i] |= value[i]
+			case XOR:
+				result[i] ^= value[i]
+			}
+		}
+	}
+	// initialize result with byteArray
+	operationResult := NewByteArray(len(result))
+	operationResult.data = result
+	operationResult.Length = int64(len(result))
+
+	// create object related to result
+	operationResultObject := store.NewObj(operationResult, -1, object.ObjTypeByteArray, object.ObjEncodingByteArray)
+
+	// store the result in destKey
+	store.Put(destKey, operationResultObject)
+
+	return makeEvalResult(len(result))
+}
