+++ internal/server/cmd_meta.go
@@ -467,6 +467,10 @@ var (
 		Cmd:     "LATENCY",
 		CmdType: SingleShard,
 	}
+	bitOPCmdMeta = CmdsMeta{
+		Cmd:     "BITOP",
+		CmdType: SingleShard,
+	}
 	// Metadata for multishard commands would go here.
 	// These commands require both breakup and gather logic.
 
@@ -599,5 +603,6 @@ func init() {
 	WorkerCmdsMeta["GEODIST"] = geodistCmdMeta
 	WorkerCmdsMeta["CLIENT"] = clientCmdMeta
 	WorkerCmdsMeta["LATENCY"] = latencyCmdMeta
+	WorkerCmdsMeta["BITOP"] = bitOPCmdMeta
 	// Additional commands (multishard, custom) can be added here as needed.
 }
